chore(*): improve tests
