+++ test/e2e/zoneegress/externalservices/e2e_suite_test.go
@@ -1,0 +1,16 @@
+package externalservices_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+
+	"github.com/kumahq/kuma/pkg/test"
+	"github.com/kumahq/kuma/test/e2e/zoneegress/externalservices"
+)
+
+func TestE2E(t *testing.T) {
+	test.RunSpecs(t, "E2E ZoneEgress for ExternalServices Suite")
+}
+
+var _ = Describe("Test ZoneEgress for External Services", externalservices.HybridUniversalGlobal)
