+++ test/framework/envoy_admin/interface.go
@@ -5,5 +5,6 @@ import (
 )
 
 type Tunnel interface {
-	GetStats(name string) (stats.Stats, error)
+	GetStats(name string) (*stats.Stats, error)
+	ResetCounters() error
 }
